// src/main/webapp/resources/js/types/irida/ExportUpoadState.ts
export enum ExportUploadState {
    NEW = "NEW",
    UPLOADING = "UPLOADING",
    UPLOADED = "UPLOADED",
    UPLOADED_ERROR = "UPLOAD_ERROR",
    CREATED = "created",
    FAILED = "failed",
    QUEUED = "queued",
    PROCESSING = "processing",
    PROCESSED_OK = "processed-ok",
    PROCESSED_ERROR = "processed-error",
    WAITING = "waiting",
    SUBMITTED = "submitted",
    DELETED = "deleted",
    RETIRED = "retired",
    UNKNOWN = "unknown",
}